requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
mssql-django==1.6
PyJWT==2.10.1
pyodbc==5.2.0
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0